refactor: unify register action helpers
